Fixed log statement for US42
- US42 now formats the date on output
- Moved format date into DateValidator from TablePrinter
-  format_date now takes the error return value
